1) Add debug define to turn off all shell iconoverlay extensions except for "synced" (for simpler debugging).
2) Construct the semaphore to shared memory and use a native pointer to then add it to the offset_ptr in the constructed Subscription.  That didn't work.
I finally found what I believe is the root problem.   The type of the offset before was size_t.  That is an unsigned type.  I changed it to uint64_t (also an unsigned type).  That worked in 32- and 64-bit processes.  However, when using the 64-bit types in a 32-bit process, arithmetic
involving pointers (unsigned 32-bit) and offsets (64-bit) does not extend the sign bit in the resulting offset.  There are two ways to change this.  1) make the offset a signed quantity (I'll try that first), and 2) add code to force-extend the sign bit wherever offsets are calculated.